Include remove and erase-all function
