+++ src/index.html
@@ -14,7 +14,7 @@
   </head>
   <body>
     <div class="box">
-      <span>Day-to-day Tasks <i class="fa-solid fa-arrows-rotate"></i></span>
+      <span>Day-to-day Tasks <i id="syncBtn" class="fa-solid fa-arrows-rotate"></i></span>
 
       <div class="insert-form">
         <input id="insertTask" type="text" placeholder="Add to your list..." />
